whitespaces, typos, more validation for IPv6 to do less service restarts

+++ home.admin/_background.sh
@@ -43,14 +43,14 @@ do
   # prevent restart if COPY OVER LAN is running
   # see: https://github.com/rootzoll/raspiblitz/issues/1179#issuecomment-646079467
   source ${infoFile}
-  if [ "${state}" == "copysource" ]; then 
+  if [ "${state}" == "copysource" ]; then
     echo "copysource mode: skipping background loop"
     sleep 10
     continue
   fi
 
   ####################################################
-  # RECHECK DHCP-SERVER 
+  # RECHECK DHCP-SERVER
   # https://github.com/rootzoll/raspiblitz/issues/160
   ####################################################
 
@@ -63,7 +63,7 @@ do
     localip=$(ip addr | grep 'state UP' -A2 | egrep -v 'docker0|veth' | grep 'eth0\|wlan0\|enp0' | tail -n1 | awk '{print $2}' | cut -f1 -d'/')
     echo "localip(${localip})"
 
-    # detect a missing DHCP config 
+    # detect a missing DHCP config
     if [ "${localip:0:4}" = "169." ]; then
       echo "Missing DHCP detected ... trying emergency reboot"
       sudo /home/admin/XXshutdown.sh reboot
@@ -127,47 +127,62 @@ do
     if [ ${publicIPChanged} -gt 0 ]; then
 
       echo "*** change of public IP detected ***"
-      echo "  old: ${publicIP}"
+
+      # store the old IP address
+      publicIP_Old="${publicIP}"
       # refresh data
       source /mnt/hdd/raspiblitz.conf
-      echo "  new: ${publicIP}"
+      # store the new IP address
+      publicIP_New="${publicIP}"
+      # some log output
+      echo "  old: ${publicIP_Old}"
+      echo "  new: ${publicIP_New}"
 
       # if we run on IPv6 only, the global IPv6 address at the current network device (e.g: eth0) is the public IP
       if [ "${ipv6}" = "on" ]; then
-        # restart bitcoind as the global IP is stored in the node configuration
-        # and we will get more connections if this matches our real IP address
-        # otherwise the bitcoin-node connections will slowly decline 
-        echo "IPv6 only is enabled => restart bitcoind to pickup up new publicIP as local IP"
-        sudo systemctl stop bitcoind
-        sleep 3
-        sudo systemctl start bitcoind
+        # if the old or the new IPv6 address is "::1" something has gone wrong in "internet.sh update-publicip" => no need to restart services
+        if [ "${publicIP_Old}" != "::1" ] && [ "${publicIP_New}" != "::1" ]; then
+          # restart bitcoind as the global IP is stored in the node configuration
+          # and we will get more connections if this matches our real IP address
+          # otherwise the bitcoin-node connections will slowly decline
+          echo "IPv6 only is enabled => restart bitcoind to pickup up new publicIP as local IP"
+          sudo systemctl stop bitcoind
+          sleep 3
+          sudo systemctl start bitcoind
 
-        # if BTCRPCexplorer is currently running 
-        # it needs to be restarted to pickup the new IP for its "Node Status Page"
-        # but this is only needed in IPv6 only mode 
-        breIsRunning=$(sudo systemctl status btc-rpc-explorer 2>/dev/null | grep -c 'active (running)')
-        if [ ${breIsRunning} -eq 1 ]; then
-          echo "BTCRPCexplorer is running => restart BTCRPCexplorer to pickup up new publicIP for the bitcoin node"
-          sudo systemctl stop btc-rpc-explorer
-          sudo systemctl start btc-rpc-explorer
-        else 
-          echo "new publicIP but no BTCRPCexplorer restart because not running"
-        fi 
-
+          # if BTCRPCexplorer is currently running
+          # it needs to be restarted to pickup the new IP for its "Node Status Page"
+          # but this is only needed in IPv6 only mode
+          breIsRunning=$(sudo systemctl status btc-rpc-explorer 2>/dev/null | grep -c 'active (running)')
+          if [ ${breIsRunning} -eq 1 ]; then
+            echo "BTCRPCexplorer is running => restart BTCRPCexplorer to pickup up new publicIP for the bitcoin node"
+            sudo systemctl stop btc-rpc-explorer
+            sudo systemctl start btc-rpc-explorer
+          else
+            echo "new publicIP but no BTCRPCexplorer restart because not running"
+          fi
+        else
+          echo "IPv6 only is ON, but publicIP_Old OR publicIP_New is equal ::1 => no need to restart bitcoind nor BTCRPCexplorer"
+        fi
       else
         echo "IPv6 only is OFF => no need to restart bitcoind nor BTCRPCexplorer"
-      fi 
+      fi
 
       # only restart LND if auto-unlock is activated
+      # AND neither the old nor the new IPv6 address is "::1"
       if [ "${autoUnlock}" = "on" ]; then
-        echo "restart LND to pickup up new publicIP"
-        sudo systemctl stop lnd
-        sudo systemctl start lnd
+        if [ "${publicIP_Old}" != "::1" ] && [ "${publicIP_New}" != "::1" ]; then
+          echo "restart LND to pickup up new publicIP"
+          sudo systemctl stop lnd
+          sudo systemctl start lnd
+        else
+          echo "publicIP_Old OR publicIP_New is equal ::1 => no need to restart LND"
+        fi
       else
         echo "new publicIP but no LND restart because no auto-unlock"
       fi
 
-      # trigger update if dnyamic domain (if set)
+      # trigger update if dynamic domain (if set)
       updateDynDomain=1
 
     else
@@ -221,7 +236,7 @@ do
     fi
     blitzTUIHeartBeatLine="${latestHeartBeatLine}"
   fi
-  
+
   ###############################
   # SCB Monitoring
   ###############################
@@ -341,7 +356,7 @@ do
   # check every hour
   recheckRAID=$((($counter % 3600)+1))
   if [ ${recheckRAID} -eq 1 ]; then
-    
+
     # check if raid is active
     source <(sudo /home/admin/config.scripts/blitz.datadrive.sh status)
     if [ ${isRaid} -eq 1 ]; then
@@ -372,7 +387,7 @@ do
 
         echo "STARTING AUTO-UNLOCK ..."
         sudo /home/admin/config.scripts/lnd.unlock.sh
-        
+
       fi
     fi
   fi
